trader item is now connected to the entity
In the base game, the trader items are added in assets/survival/itemtypes/creature.json, along with all the other creature items. You could copy that and just delete all the extra stuff not relevant to traders, that seems like the surest way to get something working. Your folder structure looks good at least, apart from the lang  folder being in entities/humanoid - that should be on the same level as entities and itemtypes.
In any case, the admin trader item should have "class": "ItemCreature", since that's the class of the item rather than the actual entity. And variantgroups might look something like this: